// packages/engine/src/syntax/diagnostic.types.ts
export enum DiagnosticSeverity {
  ERROR = 'Error',
  WARNING = 'Warning',
  INFO = 'Info',
}

export enum DiagnosticCode {
  SYNTAX_ERROR = 'SYNTAX_ERROR',
  SEMANTIC_GENERALIZATION_MISMATCH = 'SEMANTIC_GENERALIZATION_MISMATCH',
  SEMANTIC_INTERFACE_REALIZATION_INVALID = 'SEMANTIC_INTERFACE_REALIZATION_INVALID',
  SEMANTIC_CYCLE_DETECTED = 'SEMANTIC_CYCLE_DETECTED',
  SEMANTIC_DUPLICATE_MEMBER = 'SEMANTIC_DUPLICATE_MEMBER',
  SEMANTIC_INVALID_PACKAGE_RELATIONSHIP = 'SEMANTIC_INVALID_PACKAGE_RELATIONSHIP',
  SEMANTIC_INVALID_TYPE = 'SEMANTIC_INVALID_TYPE',
  SEMANTIC_IMPLICIT_ENTITY = 'SEMANTIC_IMPLICIT_ENTITY',
  SEMANTIC_INVALID_REFERENCE = 'SEMANTIC_INVALID_REFERENCE',
  SEMANTIC_INVALID_MULTIPLICITY = 'SEMANTIC_INVALID_MULTIPLICITY',
  SEMANTIC_COMPOSITE_VIOLATION = 'SEMANTIC_COMPOSITE_VIOLATION',
  SEMANTIC_DUPLICATE_ENTITY = 'SEMANTIC_DUPLICATE_ENTITY',
  SEMANTIC_AMBIGUOUS_ENTITY = 'SEMANTIC_AMBIGUOUS_ENTITY',
  SEMANTIC_INVALID_ASSOCIATION_CLASS = 'SEMANTIC_INVALID_ASSOCIATION_CLASS',
}

export interface Diagnostic {
  message: string
  code?: DiagnosticCode
  line: number
  column: number
  length?: number
  severity: DiagnosticSeverity
}